Refactor: move store files
