+++ src/app/(pages)/(club)/club/regular-club-sub/[id]/_components/CrewList.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import { useEffect, useState } from "react";
 import { getParticipationStatus, getRegularMember } from "../../../_api/supabase";
-import { useAuth } from "@/app/store/AuthContext";
+import { useAuth } from "@/store/AuthContext";
 import FullScreenModal from "./FullScreenModal";
 import NotificationList from "./NotificationList";
 import { useRouter } from "next/navigation";
@@ -147,12 +147,11 @@ const CrewList = ({ crewMembers: initialCrewMembers, clubId, clubHostId, notific
             className="w-[50%] bg-gray-100 text-primary-900"
             onClick={() => router.push(`/approvemembers/${clubId}`)}
           >
-
-              에그즈 관리
+            에그즈 관리
           </Button>
 
           <Button colorType="black" borderType="circle" sizeType="small" className="w-[50%]" onClick={handleChatClick}>
-에그클럽 채팅방
+            에그클럽 채팅방
           </Button>
         </div>
         // <div className="flex justify-center items-center gap-2">
