+++ src/components/RegularClubList.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 
 import { getRegularClubList } from "@/app/(pages)/(club)/club/_api/supabase";
 
-import { useAuth } from "@/app/store/AuthContext";
+import { useAuth } from "@/store/AuthContext";
 import { EggClubForm } from "@/types/cardlist.types";
 import { VerticalContentsListMediumEggClub } from "./uiComponents/VerticalContentsListMedium";
